requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
CTkMessagebox==1.80
customtkinter==5.1.3
cycler==0.11.0
darkdetect==0.8.0
decorator==5.1.1
fonttools==4.39.4
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pgeocode==0.4.0
Pillow==9.5.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyfiglet==0.8.post1
pyparsing==3.0.9
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2023.3
ratelim==0.1.6
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.4.1
termcolor==1.1.0
thread6==0.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
windows-curses==2.3.1
